# HackathonServer/src/main/java/com/example/HackathonServer/controllers/ChildController.java
package com.example.HackathonServer.controllers;

import com.example.HackathonServer.models.Child;
import com.example.HackathonServer.models.Move;
import com.example.HackathonServer.models.Parent;
import com.example.HackathonServer.models.Session;
import com.example.HackathonServer.repos.ChildRepo;
import com.example.HackathonServer.repos.MoveRepo;
import com.example.HackathonServer.repos.ParentRepo;
import jakarta.persistence.EntityNotFoundException;
import lombok.AllArgsConstructor;
import lombok.Data;
import lombok.NoArgsConstructor;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.http.ResponseEntity;
import org.springframework.web.bind.annotation.*;

import java.time.LocalDate;
import java.time.LocalDateTime;
import java.time.temporal.ChronoUnit;
import java.util.ArrayList;
import java.util.List;
import java.util.Map;
import java.util.Optional;

@RestController
@RequestMapping("/parent/{parentId}/child")
public class ChildController {

    @Autowired
    private ChildRepo childRepo;
    @Autowired
    private ParentRepo parentRepo;
    @Autowired
    private MoveRepo moveRepo;


    @GetMapping
    public ResponseEntity<List<Child>> getAllChildren(@PathVariable Long parentId) {
        return ResponseEntity.ok(childRepo.findByParentId(parentId));
    }

    @GetMapping("/{id}")
    public ResponseEntity<Child> getChildById(@PathVariable Long parentId, @PathVariable Long id) {
        Optional<Child> child = childRepo.findByIdAndParentId(id, parentId);
        if (child.get() == null) {
            return ResponseEntity.notFound().build();
        }
        return ResponseEntity.ok(child.get());
                
    }
    @Data
    @NoArgsConstructor
    @AllArgsConstructor

    private static class ChildRequest {
        private String fullName;
        private LocalDate dateOfBirth;
    }


    @PostMapping
    public ResponseEntity<Child> createChild(@PathVariable Long parentId, 
                                             @RequestBody ChildRequest child) {
        Parent parent = parentRepo.findById(parentId).orElse(null);
        if (parent == null) {
            return ResponseEntity.notFound().build();
        }
        Child newChild = new Child();
        newChild.setFullName(child.fullName);
        newChild.setDateOfBirth(child.dateOfBirth);
        newChild.setParent(parent);
        Child savedChild = childRepo.save(newChild);
        return ResponseEntity.ok(savedChild);
    }
    
    
    @PutMapping("/{id}")
    public ResponseEntity<Child> updateChild(@PathVariable Long parentId, @PathVariable Long id, @RequestBody Child child) {
        Optional<Child> existingChild = childRepo.findByIdAndParentId(id, parentId);
        if (existingChild.isEmpty()) {
            return ResponseEntity.notFound().build();
        }
        child.setId(id);
        child.setParent(existingChild.get().getParent());
        Child savedChild = childRepo.save(child);
        return ResponseEntity.ok(savedChild);
    }

    @DeleteMapping("/{id}")
    public ResponseEntity<Void> deleteChild(@PathVariable Long parentId, @PathVariable Long id) {
        Optional<Child> child = childRepo.findByIdAndParentId(id, parentId);
        if (child.isEmpty()) {
            return ResponseEntity.notFound().build();
        }
        childRepo.delete(child.get());
        return ResponseEntity.ok().build();
    }

    @GetMapping("/{id}/moves/last-week")
    public ResponseEntity<List<Move>> getLastWeekMoves(@PathVariable Long parentId, @PathVariable Long id) {
        Optional<Child> child = childRepo.findByIdAndParentId(id, parentId);
        if (child.isEmpty()) {
            return ResponseEntity.notFound().build();
        }
        LocalDateTime weekAgo = LocalDateTime.now().minusDays(7);
        List<Move> moves = moveRepo.findByChildIdAndCreatedAtAfter(id, weekAgo);
        return ResponseEntity.ok(moves);
    }

    @GetMapping("/{id}/moves/last-month")
    public ResponseEntity<List<Move>> getLastMonthMoves(@PathVariable Long parentId, @PathVariable Long id) {
        Optional<Child> child = childRepo.findByIdAndParentId(id, parentId);
        if (child.isEmpty()) {
            return ResponseEntity.notFound().build();
        }
        LocalDateTime monthAgo = LocalDateTime.now().minusDays(30);
        List<Move> moves = moveRepo.findByChildIdAndCreatedAtAfter(id, monthAgo);
        return ResponseEntity.ok(moves);
    }




}