# HackathonServer/src/main/java/com/example/HackathonServer/models/Child.java
package com.example.HackathonServer.models;

import com.fasterxml.jackson.annotation.JsonBackReference;
import com.fasterxml.jackson.annotation.JsonIgnore;
import com.fasterxml.jackson.annotation.JsonManagedReference;
import jakarta.persistence.*;
import lombok.AllArgsConstructor;
import lombok.Data;
import lombok.NoArgsConstructor;

import java.time.LocalDate;
import java.time.LocalDateTime;
import java.time.Period;
import java.util.List;

@Entity
@Data
@NoArgsConstructor
@AllArgsConstructor
public class Child {

    @Id
    @GeneratedValue(strategy = GenerationType.IDENTITY)
    private Long id;

    @ManyToOne
    @JoinColumn(name = "parent_id", nullable = false)
    @JsonBackReference
    private Parent parent;

    @Column(nullable = false)
    private String fullName;

    @Column(nullable = false)
    private LocalDate dateOfBirth;

    private String age;
    
    private String dp;
    @OneToMany(mappedBy = "child", cascade = CascadeType.ALL)
    @JsonManagedReference(value = "child-homework")
    @JsonIgnore
    private List<Homework> homeworks;
    private LocalDateTime createdAt;

    @OneToMany(mappedBy = "child", cascade = CascadeType.ALL)
    @JsonManagedReference(value = "child-session")
    @JsonIgnore
    private List<Session> sessions;
    private String pin;
    @PrePersist
    protected void onCreate() {
        this.createdAt = LocalDateTime.now();
        this.age = getAge();
    }
    private String getAge() {
        LocalDate currentDate = LocalDate.now();
        Period period = Period.between(dateOfBirth, currentDate);
        return period.getYears() + " y " + period.getMonths() + " m";}
}