# project/docs/apis/exergame-openapi.yaml
openapi: 3.0.3
info:
  title: Exergame API
  version: 1.0.0
servers:
  - url: http://localhost:8080
paths:
  /parent/{parentId}/child:
    get:
      tags:
        - Child
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Child"
        "400":
          description: Bad Request
        "404":
          description: Not Found
      parameters:
        - in: path
          name: parentId
          required: true
          schema:
            type: integer
            format: int64
    post:
      tags:
        - Child
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Child"
        "400":
          description: Bad Request
        "404":
          description: Not Found
      parameters:
        - in: path
          name: parentId
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: string
  /parent/{parentId}/child/{id}:
    get:
      tags:
        - Child
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Child"
        "400":
          description: Bad Request
        "404":
          description: Not Found
      parameters:
        - in: path
          name: parentId
          required: true
          schema:
            type: integer
            format: int64
        - in: path
          name: id
          required: true
          schema:
            type: integer
            format: int64
    put:
      tags:
        - Child
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Child"
        "400":
          description: Bad Request
        "404":
          description: Not Found
      parameters:
        - in: path
          name: parentId
          required: true
          schema:
            type: integer
            format: int64
        - in: path
          name: id
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Child"
    delete:
      tags:
        - Child
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: string
        "400":
          description: Bad Request
        "404":
          description: Not Found
      parameters:
        - in: path
          name: parentId
          required: true
          schema:
            type: integer
            format: int64
        - in: path
          name: id
          required: true
          schema:
            type: integer
            format: int64
  /parent/{parentId}/child/{id}/moves/last-week:
    get:
      tags:
        - Child
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Move"
        "400":
          description: Bad Request
        "404":
          description: Not Found
      parameters:
        - in: path
          name: parentId
          required: true
          schema:
            type: integer
            format: int64
        - in: path
          name: id
          required: true
          schema:
            type: integer
            format: int64
  /parent/{parentId}/child/{id}/moves/last-month:
    get:
      tags:
        - Child
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Move"
        "400":
          description: Bad Request
        "404":
          description: Not Found
      parameters:
        - in: path
          name: parentId
          required: true
          schema:
            type: integer
            format: int64
        - in: path
          name: id
          required: true
          schema:
            type: integer
            format: int64
  /homework:
    get:
      tags:
        - Homework
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Homework"
        "400":
          description: Bad Request
        "404":
          description: Not Found
    post:
      tags:
        - Homework
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Homework"
        "400":
          description: Bad Request
        "404":
          description: Not Found
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: string
  /homework/child/{childId}/all:
    get:
      tags:
        - Homework
      summary: Get all homework for a child (most recent first)
      parameters:
        - in: path
          name: childId
          required: true
          schema:
          type: integer
          format: int64
          description: ID of the child
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/HomeworkDTO"
        "404":
          description: Child not found
  /homework/{id}:
    get:
      tags:
        - Homework
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Homework"
        "400":
          description: Bad Request
        "404":
          description: Not Found
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
            format: int64
    put:
      tags:
        - Homework
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Homework"
        "400":
          description: Bad Request
        "404":
          description: Not Found
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Homework"
    delete:
      tags:
        - Homework
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: string
        "400":
          description: Bad Request
        "404":
          description: Not Found
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
            format: int64
...
components:
  schemas:
    Child:
      type: object
      properties:
        id:
          type: integer
          format: int64
        fullName:
          type: string
        dateOfBirth:
          type: string
          format: date
        age:
          type: string
        dp:
          type: string
        createdAt:
          type: string
          format: date-time
        pin:
          type: string
      required:
        - dateOfBirth
        - fullName
    Homework:
      type: object
      properties:
        id:
          type: integer
          format: int64
        title:
          type: string
        status:
          type: string
      required:
        - title
    HomeworkDTO:
      type: object
      properties:
        id:
          type: integer
          format: int64
        title:
          type: string
        url:
          type: string
        videoId:
          type: integer
          format: int64
        status:
          type: string
    Move:
      type: object
      properties:
        id:
          type: integer
          format: int64
        type:
          type: string
        createdAt:
          type: string
          format: date-time
...